Remove Legacy code in ubipop  [RHELDST-13337]
1.Legacy code in ubipop is removed.
2.ubi-manifest is the only source of expected content for ubipop.
